requirements.txt
# ==========================================
# TalaMala v4 - Dependencies
# ==========================================

# --- Web Framework ---
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
jinja2==3.1.5

# --- Database ---
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1

# --- Auth & Security ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# --- Background Jobs ---
apscheduler==3.10.4

# --- HTTP Clients ---
httpx==0.28.1
requests==2.32.3

# --- Image Processing ---
Pillow==11.1.0

# --- QR Code ---
qrcode[pil]==8.0

# --- Jalali Date ---
jdatetime==5.2.0

# --- Web Scraping (for gold price) ---
beautifulsoup4==4.12.3

# --- Environment ---
python-dotenv==1.0.1
